+++ src/gradient/typedefs.ts
@@ -9,7 +9,6 @@ export type GradientCoordValue = number | Percent | string;
 export type ColorStop = {
   color: string;
   offset: number;
-  opacity?: number;
 };
 
 export type LinearGradientCoords<T extends GradientCoordValue> = {
@@ -102,3 +101,8 @@ export type SVGOptions = {
 
   opacity: string | null;
 };
+
+export type SerializedGradientProps<T extends GradientType> = Omit<
+  GradientOptions<T>,
+  'id'
+>;